refactor: 拆分超大代码文件，确保每个文件不超过1000行
- 重构 storage.py (1160行 -> 654行 + 6个新文件)
- 重构 manager.py (1127行 -> 483行 + 8个新文件)
- 重构 config.py (1153行 -> 28行 + 5个新文件)
- 重构 auth.py (1054行 -> 7个新文件)
- 添加文件行数检查脚本
